requirement.txt
streamlit
PyPDF2
langchain
openai
faiss-cpu
python-dotenv
tiktoken